// gateway-service/src/workshop/dto/workshop-reqbody.dto.ts
export class CreateWorkshopReqBody {
  workshopTitle: string;
  workshopDetail: string;
  workshopLinkZoom: string;
  workshopPicture: string;
  workshopChefId: number;
  workshopPrice: number;
  workshopCategory: string;
  workshopIngredientPrice: number;
  workshopDate: Date;
  workshopNumberofparticipants: number;
  workshopIngredientDetail: string;
  workshopDescription: string;
}